-------------------------------------------------------    v0.20.10 (code-name) "Snitch <3"       Olvide lo que estaba haciendo con el codigo... -------------------------------------------------------    modificado:     biskana/html/html.go    modificado:     biskana/html/htmlize.go    modificado:     katana/katana.go    modificado:     katana/markup_test.go    modificado:     katana/scanner.go    modificado:     katana/table_test.go    modificado:     readme.md    modificado:     testdata/46_complex-table.morg ------------------------------------------------------- Sonic-Boom -------------------------------------------------------    Uriah Heep         : Magician's Birthday                       ! ...Very 'eavy ...Very 'umble    Gentle Giant       : Gentle Giant                       : Acquiring the Taste    Ryuichi Sakamoto   ! Music for Film    Cosmic Child       ! Blue    NieR Automata      : Piano Collections    The Rolling Stones : Stycky Fingers    Katanagatari       : Sountrack Vol. 1    Kahimi Karie       ! Larme de Crocodile                       : Montage                       ! It's Here                       : K.K. Works                       ! Specialothers                       ! Muhlifein    Sleep Party People : Lingering    BABYMETAL          ! BABYMETAL    NGExGojira         : シン・ゴジラ対エヴァンゲリオン交響楽    Roy Buchanan       ! Roy Buchanan                       ! Second Album                       : You're Not Alone                       ! That's What Am Here For                       ! When A Guitar Plays The Blues    Gary Moore         ! Back on the streets    Focus              : Focus 3    Moragami           : Noragami no Oto                       : Noragami no Oto 2    The Crystal Method ! Vegas                       ! Tweekend                       ! London                       ! Divided By Night -------------------------------------------------------

+++ katana/scanner.go
@@ -573,7 +573,7 @@ func (s *Scanner) ScanComment(ch rune) rune {
 }
 
 func (s *Scanner) SkipBlancs() {
-  for catchNewLineInScan & (1 << uint(s.Rune) ) != 0 { s.next() }
+  for catchNewLineInScan & (1 << uint64(s.Rune) ) != uint64(0) { s.next() }
 }
 
 func (s *Scanner) Scan() Token {
